src/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
httpx
apscheduler
requests
pydantic
cryptography
